+++ src/CalculoCobranca.TaxaJuro.Api/Interface/ITaxaJuroService.cs
@@ -6,6 +6,6 @@
         /// Obter o valor da taxa de juros.
         /// </summary>
         /// <returns>Retorna o valor da taxa de juros.</returns>
-        decimal ObterTaxaJuro();
+        double ObterTaxaJuro();
     }
 }
